Move type tests to the core tests directory
